refactor(web): decompose ChatPane.tsx into a features/chat-pane vertical slice
ChatPane.tsx 4,332 -> 1,194 lines. Extracted into apps/web/src/features/chat-pane/
(ports, rules.ts, 7 feature hooks, 13 dumb components, dependencies.ts) plus
apps/web/src/providers/dom.ts (8 SSR-guarded DOM bridge functions), following
docs/adr/0002-frontend-vertical-slice-decomposition.md.

Rebuilt directly against current main (not rebased from the prior unmerged
lineage) so this diffs cleanly as an isolated change: folded in main's
AMR/run-error drift since the prior attempt's fork point (chatLogTray prop,
refreshInlineAmrLoginStatus refresh-on-focus/visibilitychange, the upgrade
run-failure action plus amrPlansUrlForProfile, failureDetail-aware
resolveRunFailureUi, runFailureHasAction's primaryAction none handling,
AmrLoginPill skipInitialRefresh/onSignInStarted) into the corresponding
extracted hooks/components instead of leaving it behind.

Validation: pnpm --filter @open-design/web typecheck clean; full
@open-design/web suite green (410 files / 4,290 tests); pnpm guard clean
(78/78, including web import isolation).
