+++ apps/web/src/features/chat-pane/components/ChatConversationLoading.tsx
@@ -1,0 +1,21 @@
+import type { Dict } from '../../../i18n/types';
+
+type TranslateFn = (key: keyof Dict, vars?: Record<string, string | number>) => string;
+
+export function ChatConversationLoading({ t }: { t: TranslateFn }) {
+  return (
+    <div className="chat-loading-state" role="status" aria-live="polite">
+      <span className="chat-loading-mark" aria-hidden>
+        <span />
+        <span />
+        <span />
+      </span>
+      <span className="chat-loading-copy">{t('common.loading')}</span>
+      <span className="chat-loading-lines" aria-hidden>
+        <span />
+        <span />
+        <span />
+      </span>
+    </div>
+  );
+}
