+++ apps/web/src/features/chat-pane/components/CommentSection.tsx
@@ -1,0 +1,64 @@
+import { commentTargetDisplayName, simplePositionLabel } from '../../../comments';
+import type { PreviewComment } from '../../../types';
+
+export function CommentSection({
+  title,
+  empty,
+  comments,
+  actionLabel,
+  onAction,
+  secondaryActionLabel,
+  onSecondaryAction,
+  attached,
+}: {
+  title: string;
+  empty: string;
+  comments: PreviewComment[];
+  actionLabel: string;
+  onAction: (comment: PreviewComment) => void;
+  secondaryActionLabel?: string;
+  onSecondaryAction?: (comment: PreviewComment) => void;
+  attached?: boolean;
+}) {
+  return (
+    <section className="comments-section">
+      <h3>{title}</h3>
+      {comments.length === 0 ? (
+        <p className="comments-empty">{empty}</p>
+      ) : (
+        comments.map((comment) => (
+          <article
+            key={comment.id}
+            className={`comment-card${attached ? ' attached' : ''}`}
+            data-testid={`comment-card-${comment.elementId}`}
+          >
+            <div className="comment-card-top">
+              <strong>{commentTargetDisplayName(comment)}</strong>
+              <div className="comment-card-actions">
+                {secondaryActionLabel && onSecondaryAction ? (
+                  <button
+                    type="button"
+                    className="comment-card-action danger"
+                    onClick={() => onSecondaryAction(comment)}
+                  >
+                    {secondaryActionLabel}
+                  </button>
+                ) : null}
+                <button type="button" className="comment-card-action" onClick={() => onAction(comment)}>
+                  {actionLabel}
+                </button>
+              </div>
+            </div>
+            <p>{comment.note}</p>
+            <div className="comment-card-meta">
+              <span>{comment.id}</span>
+              <span>{comment.filePath}</span>
+              <span>{commentTargetDisplayName(comment)}</span>
+              <span>{simplePositionLabel(comment.position)}</span>
+            </div>
+          </article>
+        ))
+      )}
+    </section>
+  );
+}
