+++ apps/web/src/features/chat-pane/components/CommentsPanel.tsx
@@ -1,0 +1,56 @@
+import type { Dict } from '../../../i18n/types';
+import type { PreviewComment } from '../../../types';
+import { CommentSection } from './CommentSection';
+
+type TranslateFn = (key: keyof Dict, vars?: Record<string, string | number>) => string;
+
+export function CommentsPanel({
+  comments,
+  attachedComments,
+  onAttach,
+  onDetach,
+  onDelete,
+  t,
+}: {
+  comments: PreviewComment[];
+  attachedComments: PreviewComment[];
+  onAttach?: (comment: PreviewComment) => void;
+  onDetach?: (commentId: string) => void;
+  onDelete?: (commentId: string) => void;
+  t: TranslateFn;
+}) {
+  const attachedIds = new Set(attachedComments.map((comment) => comment.id));
+  const saved = comments.filter((comment) => !attachedIds.has(comment.id));
+  return (
+    <div className="comments-panel" data-testid="comments-panel">
+      <CommentSection
+        title={t('chat.comments.attached')}
+        empty={t('chat.comments.emptyAttached')}
+        comments={attachedComments}
+        actionLabel={t('chat.comments.remove')}
+        onAction={(comment) => onDetach?.(comment.id)}
+        attached
+      />
+      <CommentSection
+        title={t('chat.comments.saved')}
+        empty={t('chat.comments.emptySaved')}
+        comments={saved}
+        actionLabel={t('chat.comments.add')}
+        onAction={(comment) => onAttach?.(comment)}
+        secondaryActionLabel={t('chat.comments.remove')}
+        onSecondaryAction={(comment) => onDelete?.(comment.id)}
+      />
+      {saved.length > 0 ? (
+        <div className="comments-footer">
+          <button
+            type="button"
+            className="primary"
+            onClick={() => saved.forEach((comment) => onAttach?.(comment))}
+          >
+            {t('chat.comments.addAll')}
+          </button>
+        </div>
+      ) : null}
+    </div>
+  );
+}
